Added the ability in PF_RING to push filtering rule into the kernel via DAQ.
This means that when a packet received by snort is marked as bad (the whole flow and not just the packet)
PF_RING will not forward snort those packets belonging to the connection being dropped. This results in
a great performance bump of snort, that is not responsible to wasting CPU cycles with flows that PF_RING
has already stopped.